requirements.txt
streamlit==1.32.0
openai
python-dotenv
folium
streamlit-folium
requests